@@ -1,12 +1,12 @@
 apply plugin: 'com.android.library'
 apply plugin: 'kotlin-android'
-apply plugin: 'com.novoda.bintray-release'
+//apply plugin: 'com.novoda.bintray-release'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
     defaultConfig {
         minSdkVersion 14
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
     }
@@ -24,17 +24,17 @@ android {
     }
 }
 
-publish {
-    userOrg = 'ydevelop'
-    groupId = 'com.ydevelop'
-    artifactId = 'bannerlayout'
-    publishVersion = '1.1.3'
-    desc = 'this is bannerlayout'
-    website = 'https://github.com/7449/BannerLayout'
-}
+//publish {
+//    userOrg = 'ydevelop'
+//    groupId = 'com.ydevelop'
+//    artifactId = 'bannerlayout'
+//    publishVersion = '1.1.4'
+//    desc = 'this is bannerlayout'
+//    website = 'https://github.com/7449/BannerLayout'
+//}
 
 dependencies {
-    api 'com.android.support:appcompat-v7:27.1.1'
+    api 'androidx.appcompat:appcompat:1.0.0'
     compileOnly 'com.github.bumptech.glide:glide:4.8.0'
     api "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
 }
