@@ -1,9 +1,8 @@
 package com.bannersimple.simple
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import com.bannerlayout.widget.BannerLayout
-
 import com.bannersimple.R
 import com.bannersimple.bean.SimpleData
 import com.bannersimple.imagemanager.FrescoSimpleImageManager
