@@ -1,16 +1,15 @@
 package com.bannersimple.simple
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
-import android.support.v7.widget.AppCompatButton
 import android.view.View
 import android.view.Window
 import android.view.WindowManager
 import android.widget.Toast
+import androidx.appcompat.app.AppCompatActivity
+import androidx.appcompat.widget.AppCompatButton
 import com.bannerlayout.listener.OnBannerClickListener
 import com.bannerlayout.listener.SimpleOnBannerChangeListener
 import com.bannerlayout.widget.BannerLayout
-
 import com.bannersimple.R
 import com.bannersimple.bean.SimpleBannerModel
 import com.bannersimple.bean.SimpleData
