@@ -1,12 +1,11 @@
 package com.bannersimple.simple
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
-import android.support.v7.widget.AppCompatTextView
 import android.view.View
 import android.widget.AdapterView
 import android.widget.Spinner
-
+import androidx.appcompat.app.AppCompatActivity
+import androidx.appcompat.widget.AppCompatTextView
 import com.bannerlayout.listener.OnBannerChangeListener
 import com.bannerlayout.widget.BannerLayout
 import com.bannersimple.R
