@@ -3,14 +3,14 @@ package com.bannersimple.simple.issues
 import android.annotation.TargetApi
 import android.os.Build
 import android.os.Bundle
-import android.support.v4.content.ContextCompat
-import android.support.v7.app.AppCompatActivity
 import android.util.DisplayMetrics
 import android.util.Log
 import android.view.Gravity
 import android.view.View
 import android.widget.LinearLayout
 import android.widget.Toast
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.content.ContextCompat
 import com.bannerlayout.listener.SimpleOnBannerChangeListener
 import com.bannerlayout.widget.BannerLayout
 import com.bannersimple.R
